Integrate new calibration store with calibration api
This includes moving much of the tests to the integration suite
